requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyaes==1.6.1
pyasn1==0.4.8
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
schedule==1.0.0
six==1.15.0
Telethon==1.21.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
